Added fetch.js, ajax function, getCars function
Fetch.js have the DOMContentLoaded listener to execute the functions when
DOM is ready.

"ajax" function have 5 params that fetch-api use to send the request.
some params may be omitted. the success function receive
the response and error function receive the error messages.

getCars use ajax function to send the GET request and display
the response in the DOM table.